Update exclude.py
Imported the time library from python to pause the print statements after the user inputs a valid answer. Also moved the user input for their name and the subsequent print statement out of the while loop, so the program doesn't keep asking for the user's name if they answer "false".